we typically don't use 2 back to back airhorns like this. Plus generalized move messages are almost always sound 1 for GTFOs. Reserve the big emphasized sound for the attack that actually one shots you.
This however in this case, since not everyone is standingin a cloud and since it's not personalized (it's on cast not on damage or applied) it isn't a GTFO either. it's more of a dodge/watch your step

This leaves the most accurate warning type to be a type 2 aoe dodge message. Warning level 2 also isn't already used on this fight either which further diversifies non voiced sound selection.

So with all above in mind, set Deep Slumber to an aoe dodge ("watch your step")  warning