Add a "block" for Gtfo warnings
Blocks are a new experimental thing to reduce repetitive code for recurring patterns in mods. I'll be using the Geddon mod as a first example as it has two common patterns: Gtfo warnings and bombs on players.

As a side-effect this fixes a bug about a Gtfo warning not working properly in SoD